+++ Arbiter.App/ViewModels/Dialogs/DialogMenuChoiceViewModel.cs
@@ -6,7 +6,11 @@ public partial class DialogMenuChoiceViewModel : ViewModelBase
 {
     [ObservableProperty] private string _text = string.Empty;
 
-    [ObservableProperty] private int? _pursuitId;
+    [ObservableProperty] 
+    [NotifyPropertyChangedFor(nameof(IsModMenuChoice))]
+    private int? _pursuitId;
 
     public required int Index { get; init; }
+
+    public bool IsModMenuChoice => PursuitId >= 0xFF00;
 }