+++ Arbiter.App/ViewModels/Proxy/ProxyViewModel.DialogFilters.cs
@@ -1,9 +1,11 @@
 ﻿using System;
+using System.Linq;
 using Arbiter.App.Models;
 using Arbiter.Net;
 using Arbiter.Net.Filters;
 using Arbiter.Net.Proxy;
 using Arbiter.Net.Server.Messages;
+using Arbiter.Net.Types;
 
 namespace Arbiter.App.ViewModels.Proxy;
 
@@ -11,21 +13,26 @@ public partial class ProxyViewModel
 {
     private NetworkFilterRef? _debugDialogFilter;
     private NetworkFilterRef? _debugDialogMenuFilter;
+    private NetworkFilterRef? _debugDialogMenuItemQuantityFilter;
 
     private void AddDebugDialogFilters(DebugSettings settings)
     {
         _debugDialogFilter = _proxyServer.AddFilter<ServerShowDialogMessage>(HandleDialogMessage,
             $"{FilterPrefix}_Dialog_ServerShowDialog", DebugFilterPriority, settings);
 
-
         _debugDialogMenuFilter = _proxyServer.AddFilter<ServerShowDialogMenuMessage>(HandleDialogMenuMessage,
             $"{FilterPrefix}_Dialog_ServerShowDialogMenu", DebugFilterPriority, settings);
+
+        _debugDialogMenuItemQuantityFilter = _proxyServer.AddFilter<ServerShowDialogMenuMessage>(
+            HandleDialogItemQuantityMenuMessage,
+            $"{FilterPrefix}_Dialog_ItemQuantity_ServerShowDialogMenu", int.MinValue, settings);
     }
 
     private void RemoveDebugDialogFilters()
     {
         _debugDialogFilter?.Unregister();
         _debugDialogMenuFilter?.Unregister();
+        _debugDialogMenuItemQuantityFilter?.Unregister();
     }
 
     private static NetworkPacket HandleDialogMessage(ProxyConnection connection, ServerShowDialogMessage message,
@@ -92,7 +99,6 @@ public partial class ProxyViewModel
                     var choiceText = choice.Text.Length > maxChoiceLength
                         ? string.Concat(choice.Text.AsSpan(0, maxChoiceLength - 3), "...")
                         : choice.Text;
-                    ;
 
                     choice.Text = $"{choiceText} {menuPursuitText}";
                 }
@@ -102,5 +108,59 @@ public partial class ProxyViewModel
         }
 
         return hasChanges ? result.Replace(message) : result.Passthrough();
+    }
+
+    private NetworkPacket HandleDialogItemQuantityMenuMessage(ProxyConnection connection,
+        ServerShowDialogMenuMessage message, object? parameter,
+        NetworkMessageFilterResult<ServerShowDialogMenuMessage> result)
+    {
+        if (parameter is not DebugSettings settings || settings is { ShowDialogItemQuantity: false } ||
+            message.MenuType != DialogMenuType.UserInventory)
+        {
+            return result.Passthrough();
+        }
+
+        // If unable to get the player OR no stackable items just passthrough
+        if (!_playerService.TryGetState(connection.Id, out var player) || !player.Inventory.Any(i => i.IsStackable))
+        {
+            return result.Passthrough();
+        }
+        
+        SendItemQuantityOverrides(connection, player.Inventory);
+        return result.Passthrough();
+    }
+
+    private static void SendItemQuantityOverrides(ProxyConnection connection, PlayerInventory inventory)
+    {
+        // First we update the client with the stack size in the name
+        var addItemMessages = inventory.Where(i => i.IsStackable)
+            .Select(item => new ServerAddItemMessage
+            {
+                Slot = (byte)item.Slot,
+                Name = $"{item.Name} [{item.Quantity}]",
+                Sprite = item.Sprite,
+                Color = (DyeColor)item.Color,
+                Quantity = (uint)item.Quantity,
+                IsStackable = item.IsStackable,
+                Durability = (uint)(item.Durability ?? 0),
+                MaxDurability = (uint)(item.MaxDurability ?? 0)
+            });
+        connection.EnqueueMessages(addItemMessages, NetworkPriority.High);
+        
+        // Then we revert it after a short delay which will happen after the dialog is shown
+        // This happens very quickly so its not noticeable to the user
+        var revertItemMessages = inventory.Where(i => i.IsStackable)
+            .Select(item => new ServerAddItemMessage
+            {
+                Slot = (byte)item.Slot,
+                Name = item.Name,
+                Sprite = item.Sprite,
+                Color = (DyeColor)item.Color,
+                Quantity = (uint)item.Quantity,
+                IsStackable = item.IsStackable,
+                Durability = (uint)(item.Durability ?? 0),
+                MaxDurability = (uint)(item.MaxDurability ?? 0)
+            });
+        connection.EnqueueMessagesAfter(revertItemMessages, TimeSpan.FromMilliseconds(100), NetworkPriority.High);
     }
 }