+++ Arbiter.App/ViewModels/Proxy/ProxyViewModel.ModMenuFilters.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
 using Arbiter.App.Models;
 using Arbiter.Net;
 using Arbiter.Net.Client.Messages;
@@ -12,6 +13,8 @@ namespace Arbiter.App.ViewModels.Proxy;
 
 public partial class ProxyViewModel
 {
+    private const int ModMenuFilterPriority = int.MinValue + 100;
+    
     private static readonly List<(string Text, ushort PursuitId)> ModMenuChoices =
     [
         ("Interact", 0xFF00),
@@ -33,10 +36,10 @@ public partial class ProxyViewModel
     private void AddModMenuFilters(DebugSettings settings)
     {
         _modMenuInteractFilter = _proxyServer.AddFilter<ClientInteractMessage>(HandleModMenuInteractMessage,
-            $"{FilterPrefix}_ModMenu_ClientInteract", int.MinValue, settings);
+            $"{FilterPrefix}_ModMenu_ClientInteract", ModMenuFilterPriority, settings);
 
         _modMenuInteractFilter = _proxyServer.AddFilter<ClientDialogMenuChoiceMessage>(HandleModMenuDialogChoiceMessage,
-            $"{FilterPrefix}_ModMenu_ClientInteract", int.MinValue, settings);
+            $"{FilterPrefix}_ModMenu_ClientInteract", ModMenuFilterPriority, settings);
     }
 
     private void RemoveModMenuFilters()
@@ -97,7 +100,13 @@ public partial class ProxyViewModel
         if (message.PursuitId == 0xFF01 && _entityStore.TryGetEntity(message.EntityId, out var entity) &&
             entity.Flags.HasFlag(EntityFlags.Mundane))
         {
-            var destroyItemMenu = GetDestroyItemDialogForEntity(entity);
+            // Get the player so we can get their inventory
+            _playerService.TryGetState(connection.Id, out var player);
+            var inventory = player?.Inventory ?? new PlayerInventory();
+            
+            var destroyItemMenu = GetDestroyItemDialogForEntity(entity, inventory);
+            
+            SendItemQuantityOverrides(connection, inventory);
             connection.EnqueueMessage(destroyItemMenu);
             return result.Block();
         }
@@ -139,7 +148,8 @@ public partial class ProxyViewModel
         return dialog;
     }
 
-    private static ServerShowDialogMenuMessage GetDestroyItemDialogForEntity(GameEntity entity)
+    private static ServerShowDialogMenuMessage GetDestroyItemDialogForEntity(GameEntity entity,
+        PlayerInventory inventory)
     {
         var dialog = new ServerShowDialogMenuMessage
         {
@@ -159,12 +169,9 @@ public partial class ProxyViewModel
             Content =
                 "Which item do you wish to destroy?\n{=hThis may not work on items which cannot be dropped.\n{=sBe careful! This cannot be undone.\n"
         };
-
-        // Add all inventory slots
-        for (var i = 1; i < 60; i++)
-        {
-            dialog.InventorySlots.Add((byte)i);
-        }
+        
+        // Add all used slots to the dialog
+        dialog.InventorySlots.AddRange(inventory.GetNonEmptySlots().Select(x => (byte)x));
 
         return dialog;
     }