+++ Arbiter.App/ViewModels/Proxy/ProxyViewModel.cs
@@ -12,8 +12,9 @@ public partial class ProxyViewModel : ViewModelBase
 {
     private readonly ILogger<ProxyViewModel> _logger;
     private readonly ProxyServer _proxyServer;
+    private readonly IPlayerService _playerService;
     private readonly IEntityStore _entityStore;
-
+    
     public bool IsRunning => _proxyServer.IsRunning;
 
     public ProxyViewModel(ILogger<ProxyViewModel> logger, IServiceProvider serviceProvider)
@@ -21,6 +22,7 @@ public partial class ProxyViewModel : ViewModelBase
         _logger = logger;
         
         _proxyServer = serviceProvider.GetRequiredService<ProxyServer>();
+        _playerService = serviceProvider.GetRequiredService<IPlayerService>();
         _entityStore = serviceProvider.GetRequiredService<IEntityStore>();
     }