+++ Arbiter.App/Views/PlayerInventoryView.axaml.cs
@@ -1,0 +1,11 @@
+﻿using Avalonia.Controls;
+
+namespace Arbiter.App.Views;
+
+public partial class PlayerInventoryView : UserControl
+{
+    public PlayerInventoryView()
+    {
+        InitializeComponent();
+    }
+}