+++ Arbiter.Net/NetworkPacketSource.cs
@@ -1,0 +1,7 @@
+﻿namespace Arbiter.Net;
+
+public enum NetworkPacketSource
+{
+    Network = 0,
+    Injected = 1
+}